+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Braintree Android Drop-In Release Notes
 
+## unreleased
+* Visibility
+  * Make `BottomSheetState` package-private
+  * Make `DropInEventType` package-private
+  * Make `DropInActivity#onError` package-private
+  * Make `AnimatedButtonView` package-private
+  * Make `PaymentMethodItemView` package-private
+  * Make `SupportedPaymentMethodsViewHolder` package-private
+  * Make `VaultedPaymentMethodSelectedListener` package-private
+
 ## 6.0.0-beta2
 
 * Android 12 Support